Port caml-types-time to future Emacs
* caml-types.el (caml-types-time):
Do not assume Emacs timestamp format, which is slated to change in
future versions.  Instead, use time-convert to convert to milliseconds
since the epoch modulo 1000000.  This is simpler and less glitchy than
the old version anyway (millseconds since the epoch modulo 1000, plus
1000 times seconds since the epoch modulo 65536 modulo 1000).
This code uses fboundp to be portable to Emacs 26 and earlier,
and can be simplified once we can assume Emacs 27.